add update job for upload SR and DD
