+++ app/airflow/dags/libs/enums.py
@@ -1,6 +1,20 @@
-from enum import StrEnum
+from enum import StrEnum, Enum
 
 
 class StorageType(StrEnum):
     AZURE = "azure"
     MINIO = "minio"
+
+
+class StageStatusType(Enum):
+    IN_PROGRESS = "Job in Progress"
+    COMPLETE = "Job Complete"
+    FAILED = "Job Failed"
+
+
+class JobStageType(Enum):
+    UPLOAD_SCAN_REPORT = "Upload Scan Report"
+    BUILD_CONCEPTS_FROM_DICT = "Build concepts from OMOP Data dictionary"
+    REUSE_CONCEPTS = "Reuse concepts from other scan reports"
+    GENERATE_RULES = "Generate mapping rules from available concepts"
+    DOWNLOAD_RULES = "Generate and download mapping rules JSON"
